.history/backend/requirements_20250527224449.txt
fastapi==0.109.2
uvicorn==0.27.1
python-socketio==5.11.1
google-cloud-speech==2.24.1
python-multipart==0.0.9
websockets==12.0
eventlet==0.35.2
google-generativeai 